Ignore undefined terms when expanding
Updated the expansion algorithm to ignore undefined terms that are not in the form of a compact IRI or absolute IRI. Removed wrong "Otherwise" that implied that the action gets just executed if the previous ones haven't - which isn't true. Furthermore, reverted removal of active property parameter from expansion algorithm done in 3aa65a7ef8f7216b7276427bdd61cf0e8f887e5a as that was a mistake.