Update subject map generation algorithm to fix bugs and handle multiple graphs
The current subject map generation algorithm didn't handle multiple graphs and had some bugs. This update includes full support for @graph and fixes those bugs. If the values of @type would be expanded to @id objects (issue #118), this algorithm could be further simplified.

This addresses #124. Since the bug is also in the playground I do not close this issue yet.