Define default values of JsonLdOptions
Unfortunately the used version of ReSpec (although the test at [1] support it) doesn't handle default values correctly so that I had to add them as well in prose.

Furthermore the formatting is wrong: Instead of using h4 headers respec uses h2 headers for WebIDL subsections. This has been fixed in [2] but I think we are using a different version.

[1] http://dev.w3.org/2009/dap/ReSpec.js/test-spec/webidl.html#with-members
[2] https://github.com/darobin/respec/issues/46

This closes #112.