Rename "nested contexts" to "scoped contexts" and remove External Contexts section
Almost all the information contained in the External Contexts section was already contained the The Context section. The only thing missing there was the scoped context example which was moved as part of this commit. Therefore I also completely removed the External Context section at the same time.

This closes #66.