Re-sort keywords in spec and lang-jsonld.js
They are now in the same order as they are introduced in the spec. The only exception is @vocab which is introduced in the IRIs section but should be considered a more advanced feature than @id or @value for instance.

There's a discussion about this ordering in abcf5ae69823e39ccf0489f18234f0b41d4b31e3.