New "stand-alone" Grammar section
Main changes:
- Clarified that JSON terms are directly imported from the JSON spec,
and not re-defined in JSON-LD.
- Moved the definitions of "node definition", "node reference" and
"term" to the grammar. They are JSON-LD constructs.
- Introduced new grammar constructs when needed, in particular
"expanded typed value" and "expanded lanugage-tagged string"
- Re-wrote the grammar in a more consistent way to use the same
vocabulary throughout.
- Fixed grammar when needed (for instance the value in a language
map could be a language-tagged string, which seemed a typo).

Not done yet:
- All normative statements in "Basic Concepts" and "Advanced Concepts"
should be removed (provided they appear in the grammar that is!)
- There remains a couple of places where the grammar is not perfect
(@graph and apparent contradiction in expanded term definition in
particular)
- Examples in the grammar should be moved to previous sections.
- There may be a few places in "Basic Concepts" and "Advanced Concepts"
sections that would benefit from linking to the grammar.
- Some definitions may benefit from a bit of wordsmithing. For instance,
there could be a note around the definition of an IRI that explains
that, whenever the grammar says a value must be an IRI, it is taken
to mean "a string with the lexical form of an IRI".