Dropped normative statement from "Basic" and "Advanced" sections.
Grammar completed as necessary when the normative statements were
not captured yet (e.g. the constraint on circular dependencies
for terms definitions in a context).

Also: IRI definition modified to say that an IRI may either be
represented as an absolute IRI or a relative IRI. "terms" and
"compact IRIs" were also in the list previously, but they generate
IRIs through expansion, and the grammar makes the distinction
when it says that a value can be an IRI, a compact IRI or a term.

Normative statements on JSON-LD processors still appear in the
doc, but should probably be removed.

Also note the "Referencing Contexts from JSON Documents" section
retains its normative wording.

I also updated plural references to terms to include the "s" to
ensure the whole world gets turned into a link.