Removed definition of JSON-LD processors from the Conformance section
Other sections still mention JSON-LD processors, either to describe
how JSON-LD documents are processed or to put constraints on JSON-LD
processors through normative statements. Descriptive text would need
to link to the proper definition of JSON-LD processor in the API doc.
Normative statements that apply to JSON-LD processors need to be
dropped. All of this is being covered by issue #184.