Examples in the grammar section moved to other sections or dropped
As agreed in #114, I moved all the examples that appeared in the
grammar section to the "Basic Concepts" and "Advanced Concepts"
sections when appropriate.

I dropped examples that were already duplicates of examples in these
sections as well as examples that were very close to other examples.

The grammar section does not directly link to the examples in the
other sections but links to relevant sections each time. That seems
enough to me (and that also simplifies things since ReSpec does not
naturally create IDs for examples, making it a pain to link to them
directly).