Remove contradictory statement from grammar
The statement was:

"An expanded value that contains a @type key is called an expanded typed value. An expanded value that contains a @language key is called an expanded language-tagged string. An expanded value that contains neither the @type key or the @language key **is both** an expanded typed value and an expanded language-tagged string."