Replace index.html with alternate2.html except examples in intro
I kept the examples that use FOAF and for the time being the old flattening example. I adjusted the indentation of the merged text so that it's easier to see what stayed there unchanged. I also kept the "conflicting indexes" error code instead of "CONFLICTING_ANNOTATIO" as in alternate2.html, and the document head. All the rest has been merged as is was in alternate2.html.

I renamed the old index.html to markus.html.