Make naming of algorithms more consistent (2/2)
This is probably controversial, thus the separate commit so that we can easily revert it. I think almost all algorithms are "subalgorithms". Just marking a few as such doesn't improve clarity. The approach I'm taking here is to add the word "Algorithm" to the main algorithms and drop it (as well as "Subalgorithm") from all others.

//cc @dlongley @gkellogg

This addresses #218.