Some minor fixes and improvements to Grammar
Mostly things were reordered. There were two notable changes:

I made was to rename "expanded value" to "value object" to be in line with the API spec. I also introduced the notion of "list object" and "set object" instead of just calling them sets and lists (defined in the data model).

I changed the value space of @base to the be an absolute IRI or null (before it also included relative IRIs) (see also #223).