Use i-tag to enclose variables instead of em
Using a span with a class is a lot of work when typing. I think the i-tag does the job. Since it's not used for anything else, it is also easy to replace it with something else if needed.

For the time being I didn't change the style of variables - I couldn't find anything that looks good :-)

This addresses #218.