Combine Convert to RDF Algorithm and Node Map to RDF Conversion
Both algorithms are short and there are no recursions. It's easier to read them when they are together instead of being in separate sections. I've also clarified Object to RDF Conversion and List to RDF Conversion a bit.

Perhaps also the List to RDF Conversion should be folded into the Convert to RDF Algorithm!?

/cc @gkellogg

This addresses #218.