Pub rules checker doesn't like scheme-relative URL, convert to absolute
I don't know how strictly such things are treated but a scheme-relative URL would be much better in this instance because it would work on both https and http. Otherwise, in some browser the stylesheet is not processed when the document is served over https (e.g. https://dvcs.w3.org)