API edits to avoid implying that JSON-LD is not RDF
As promised, attached are suggested edits to the "JSON-LD 1.0 Processing
Algorithms and API" document at

  http://json-ld.org/spec/latest/json-ld-api/index.html

The purpose of these edits is to avoid implying that JSON-LD is not RDF.

The main strategy was to change phrases like "transforms JSON-LD to
RDF" and "converts RDF to JSON-LD" to phrases like "deserializes JSON-LD
to RDF" and "serializes RDF to JSON-LD".

Note that some of the section titles changed as a result, and two HTML
anchors were also changed:

  convert-to-rdf-algorithm --> deserialize-to-rdf-algorithm

  convert-from-rdf-algorithm --> serialize-from-rdf-algorithm

Therefore other documents that reference this document may need to
be updated correspondingly.