Syntax doc edits to avoid implying that JSON-LD is not RDF
Attached are the corresponding changes that I see that are needed in the
syntax document: http://json-ld.org/spec/latest/json-ld/

Other editorial issues that I noticed but I did not fix in the attached
version:

1. The Abstract begins: "JSON is a useful data serialization and
messaging format".  Delete "useful" or perhaps delete the whole
sentence.  (Other data serialization and messaging formats are *not*
useful?)

2. The Features at Risk section needs to be updated regarding blank
nodes as predicates.