Add playground context map support.
Support was added for the playground to use a documentLoader that has a
mapping between specific user specified contexts and a modified context
to load. Currently this is used to map the use of the non-existent
schema.org context to a temporary at w3id.org and hosted at json-ld.org.

A bit of UI feedback is provided to let users know which contexts were
mapped to modified values.