Fixed duplicate json keys in expand-0004 unit test files.
Most implementations of json parsers seem to use dictionaries (hash tables) as representation of json objects. Because of this, when duplicate json key appears while parsing, the value for the last appearance of the key is used in the output dictionary. I found the error in unit test files while working with a parser implementation that uses lists of pairs to represent json objects.