Initial Bootstrap 5 wrapping template.
Really not a fan of the Bootstrap 5 UI kit. It is very low contrast
(bad for accessibility) and changes most of the old Bootstrap classes
(making upgrade non-trivial).

Consequently, any other CSS framework would be nearly as complex to
upgrade to...so exploring other options is back on the table.