Add Table output.
This mostly matches the current implementation with one notable exception:
it outputs the datatype of XSD Strings or RDF langString values (which were
previously hidden). It is not clear to me why hiding those would be useful
and instead would seem to make the output in the table less accurate.

We can hide those later if people feel there is value in doing that.